requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
face-recognition==1.3.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
torch==2.0.1
torchvision==0.15.2
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
tqdm==4.66.1 